requirements.txt
dj-database-url==0.5.0
Django==1.11.18
django-heroku==0.3.1
gunicorn==19.9.0
jsonfield==1.0.3
Pillow==3.4.0
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
djangorestframework
djangorestframework-jwt
